+++ src/components/AddExpensePage.js
@@ -1,0 +1,9 @@
+import React from 'react'
+
+const AddExpensePage = () => (
+  <div>
+    this is from the addExpense
+  </div>
+)
+
+export default AddExpensePage
