+++ src/components/ExpenseDashboardPage.js
@@ -1,0 +1,9 @@
+import React from 'react'
+
+const  ExpenseDashboardPage = () => (
+  <div>
+    this is from the ExpenseDashboardPage
+  </div>
+)
+
+export default ExpenseDashboardPage
