+++ src/components/HelpPage.js
@@ -1,0 +1,9 @@
+import React from 'react'
+
+const HelpPage = () => (
+  <div>
+    this is from the HelpPage
+  </div>
+)
+
+export default HelpPage
