+++ src/routers/AppRouter.js
@@ -1,0 +1,26 @@
+import React from 'react'
+import { BrowserRouter, Route, Switch } from 'react-router-dom'
+import AddExpensePage from '../components/AddExpensePage'
+import EditExpensePage from '../components/EditExpensePage'
+import ExpenseDashboardPage from '../components/ExpenseDashboardPage'
+import Header from '../components/Header'
+import HelpPage from '../components/HelpPage'
+import NotFoundPage from '../components/NotFoundPage'
+
+const AppRouter = () => (
+  <BrowserRouter>
+    <div>
+      <Header />
+      <Switch>
+        <Route path="/" component={ExpenseDashboardPage} exact={true}/>
+        <Route path="/create" component={AddExpensePage} />
+        <Route path="/edit/:id" component={EditExpensePage} />
+        <Route path="/help" component={HelpPage} />
+        <Route component={NotFoundPage} />
+      </Switch>
+    </div>
+
+  </BrowserRouter>
+)
+
+export default AppRouter
